// src/interfaces/brewerie.interface.ts
export interface IBrewerieModel {
    id: number;
    obdbId: string | undefined;
    name: string | undefined;
    breweryType: string | undefined;
    street: string | undefined;
    address2: string | undefined;
    address3: string | undefined;
    city: string | undefined;
    state: string | undefined;
    countyProvince: string | undefined;
    postalCode: string | undefined;
    country: string | undefined;
    longitude: string | undefined;
    latitude: string | undefined;
    phone: string | undefined;
    websiteUrl: string | undefined;
    updatedAt: string;
    createdAt: string;
    region: string | undefined;
}